More linting and update to 1.19 [#3146]
Closes #3146. Updates to 1.19 and adds linting.

- Updates to 1.19, which adds new formatting rules for documentation comments.
- Adds more lint rules, some suggested by Notional, some chosen by me.

Changelog: refactor

Former-commit-id: b2f3160813abba55aca4d5383b8d4609c3fcddbc

+++ internal/database/smt/pmt/bpt_savestate.go
@@ -157,9 +157,9 @@ func ReadSnapshot(file ioutil2.SectionReader, storeState func(key storage.Key, h
 	}
 	numNodes, _ := common.BytesFixedUint64(vBuff) //          Convert entry count to number of BPT nodes
 
-	fOff := uint64(8 + numNodes*(nLen)) //                    File offset to values in snapshot
-	fileIndex := int64(8)               //                    Used to track progress through the file
-	toRead := window                    //                    Assume a window's worth to read
+	fOff := 8 + numNodes*(nLen) //                    File offset to values in snapshot
+	fileIndex := int64(8)       //                    Used to track progress through the file
+	toRead := window            //                    Assume a window's worth to read
 	for toRead > 0 {
 
 		if numNodes < toRead { //                             If not a window's worth of nodes left
